v 0.2.0 FIX `make env` and setup.py
setup.py was previously importing geco_stat, which caused problems when
initializing the environment since the dependencies are not installed in
a fresh virtualenv. fixed this by putting version information in a separate
version.py file, from which both geco_stat.py and setup.py are able to get
the latest version information.

also, changed the names VERSION and RELEASE in geco_stat to __version__
and __release__, and tweak Makefile.