FIX Read The Docs build dependency problems
FIX RTD can't build extensions that rely on hard c code, like `h5py`. Fix this
by creating mock modules for these tests (which is not such a bad idea
anyway for the sake of unit testing and should be implemented elsewhere
when I find the time). As part of this fix, create a separate, empty
`docs/requirements.rtd.txt` file, which must be explicitly selected as
the RTD requirements file on the Read The Docs project page.

ADD the RTD theme to doc builds initiated on this machine. Makes it easier to
see what the final product will look like.