Ignore too-many-arguments landscape.io
There are several abstract classes that take long lists of arguments as part of their constructors. This allows for flexible and clear interfaces and should not be changed, in particular because these methods should not be commonly used for scripting or interactive sessions anyway and should therefore not interfere with usability. Landscape.io should hence ignore this type of error.